% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-regression.R
\name{coef.congruence_spline}
\alias{coef.congruence_spline}
\title{Extract congruence spline coefficients}
\usage{
\method{coef}{congruence_spline}(object, ...)
}
\arguments{
\item{object}{A \code{congruence_spline} object.}

\item{...}{Unused.}
}
\value{
A named numeric vector of model coefficients.
}
\description{
Extract congruence spline coefficients
}
