% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-regression.R
\name{nobs.congruence_spline}
\alias{nobs.congruence_spline}
\title{Number of observations in a fitted congruence spline}
\usage{
\method{nobs}{congruence_spline}(object, ...)
}
\arguments{
\item{object}{A \code{congruence_spline} object.}

\item{...}{Unused.}
}
\value{
The number of complete cases used to fit the model.
}
\description{
Number of observations in a fitted congruence spline
}
