% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-regression.R
\name{print.congruence_spline}
\alias{print.congruence_spline}
\title{Print a congruence spline model}
\usage{
\method{print}{congruence_spline}(x, ...)
}
\arguments{
\item{x}{A \code{congruence_spline} object.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a congruence spline model
}
