% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-regression.R
\name{vcov.congruence_spline}
\alias{vcov.congruence_spline}
\title{Extract a congruence spline covariance matrix}
\usage{
\method{vcov}{congruence_spline}(object, ...)
}
\arguments{
\item{object}{A \code{congruence_spline} object.}

\item{...}{Unused.}
}
\value{
A numeric covariance matrix. Values are \code{NA} if \code{numDeriv} is not
installed.
}
\description{
Extract a congruence spline covariance matrix
}
