% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsahelpers-package.R
\docType{package}
\name{rsahelpers-package}
\alias{rsahelpers}
\alias{rsahelpers-package}
\title{rsahelpers: Helpers for Response Surface Analysis}
\description{
\code{rsahelpers} provides tools for fitting and inspecting response surfaces,
including Edwards-Parry spline regression models and Mplus coefficient
extraction and plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franciscowilhelm/rsahelpers}
  \item \url{https://franciscowilhelm.github.io/rsahelpers/}
  \item Report bugs at \url{https://github.com/franciscowilhelm/rsahelpers/issues}
}

}
\author{
\strong{Maintainer}: Francisco Wilhelm \email{francisco.wilhelm@unibe.ch} (\href{https://orcid.org/0000-0002-1717-7975}{ORCID})

Authors:
\itemize{
  \item Francisco Wilhelm \email{francisco.wilhelm@unibe.ch} (\href{https://orcid.org/0000-0002-1717-7975}{ORCID})
}

}
\keyword{internal}
